Switch to the jupyterlab test
