+++ tests/e2e/jest-puppeteer.config.js
@@ -3,7 +3,7 @@ const demo = false;
 module.exports = {
   launch: {
     headless: !demo,
-    slowMo: demo ? 60 : 0,
+    slowMo: demo ? 10 : 0,
     defaultViewport: {
       width: 1440,
       height: 900
