+++ tests/e2e/tutorials/tutorialBase.js
@@ -244,7 +244,7 @@ class TutorialBase {
     return resp;
   }
 
-  async waitForServices(studyId, nodeIds, timeout = 40000) {
+  async waitForServices(studyId, nodeIds, timeout = 40000, waitForConnected = true) {
     if (nodeIds.length < 1) {
       return;
     }
@@ -253,7 +253,12 @@ class TutorialBase {
     while ((new Date().getTime()) - start < timeout) {
       for (let i = nodeIds.length - 1; i >= 0; i--) {
         const nodeId = nodeIds[i];
-        if (await utils.isServiceReady(this.__page, studyId, nodeId) && await utils.isServiceConnected(this.__page, studyId, nodeId)) {
+        let isLoaded = await utils.isServiceReady(this.__page, studyId, nodeId);
+        if(waitForConnected) {
+          isLoaded = isLoaded && await utils.isServiceConnected(this.__page, studyId, nodeId);
+        }
+
+        if (isLoaded) {
           nodeIds.splice(i, 1);
         }
       }
@@ -411,12 +416,12 @@ class TutorialBase {
     await this.takeScreenshot("closeStudy_after");
   }
 
-  async removeStudy(studyId) {
-    await this.waitFor(5000, 'Wait to be unlocked');
+  async removeStudy(studyId, timeout=5000) {
+    await this.waitFor(timeout, 'Wait to be unlocked');
     await this.takeScreenshot("deleteFirstStudy_before");
     try {
       // await this.waitForStudyUnlocked(studyId);
-      const nTries = 3;
+      const nTries = 30;
       let i
       for (i = 0; i < nTries; i++) {
         const cardUnlocked = await auto.deleteFirstStudy(this.__page, this.__templateName);
